+++ circus-train-core/src/main/java/com/hotels/bdp/circustrain/core/replica/FullReplicationReplicaLocationManager.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2018 Expedia Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,27 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.hotels.bdp.circustrain.core.replica;
 
 import static com.hotels.bdp.circustrain.core.replica.TableType.UNPARTITIONED;
 import static com.hotels.hcommon.hive.metastore.util.LocationUtils.locationAsPath;
-
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.List;
 
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hive.metastore.api.Partition;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.common.annotations.VisibleForTesting;
-
 import com.hotels.bdp.circustrain.api.ReplicaLocationManager;
 import com.hotels.bdp.circustrain.api.SourceLocationManager;
 import com.hotels.bdp.circustrain.api.event.ReplicaCatalogListener;
-import com.hotels.bdp.circustrain.api.listener.HousekeepingListener;
 
 public class FullReplicationReplicaLocationManager implements ReplicaLocationManager {
 
@@ -42,33 +34,25 @@ public class FullReplicationReplicaLocationManager implements ReplicaLocationMan
   private final SourceLocationManager sourceLocationManager;
   private final String tablePath;
   private final String eventId;
-  private final List<CleanUpLocation> previousLocations;
   private final TableType tableType;
-  private final HousekeepingListener housekeepingListener;
+
+  private final CleanupLocationManager cleanupLocationManager;
+
   private final ReplicaCatalogListener replicaCatalogListener;
-
-  private final String replicaDatabaseName;
-
-  private final String replicaTableName;
 
   FullReplicationReplicaLocationManager(
       SourceLocationManager sourceLocationManager,
       String tablePath,
       String eventId,
       TableType tableType,
-      HousekeepingListener housekeepingListener,
-      ReplicaCatalogListener replicaCatalogListener,
-      String replicaDatabaseName,
-      String replicaTableName) {
+      CleanupLocationManager cleanupLocationManager,
+      ReplicaCatalogListener replicaCatalogListener) {
     this.sourceLocationManager = sourceLocationManager;
     this.tablePath = tablePath;
     this.eventId = eventId;
     this.tableType = tableType;
+    this.cleanupLocationManager = cleanupLocationManager;
     this.replicaCatalogListener = replicaCatalogListener;
-    this.replicaDatabaseName = replicaDatabaseName;
-    this.replicaTableName = replicaTableName;
-    previousLocations = new ArrayList<>();
-    this.housekeepingListener = housekeepingListener;
   }
 
   @Override
@@ -94,21 +78,12 @@ public class FullReplicationReplicaLocationManager implements ReplicaLocationMan
 
   @Override
   public void addCleanUpLocation(String pathEventId, Path location) {
-    LOG.debug("Adding clean up location: {}", location.toUri());
-    previousLocations.add(new CleanUpLocation(pathEventId, location));
+    cleanupLocationManager.addCleanupLocation(pathEventId, location);
   }
 
   @Override
   public void cleanUpLocations() {
-    List<URI> uris = new ArrayList<>();
-    for (CleanUpLocation location : previousLocations) {
-      LOG.info("Scheduling old replica data for deletion for event {}: {}", eventId, location.getPath().toUri());
-      housekeepingListener
-          .cleanUpLocation(eventId, location.getPathEventId(), location.getPath(), replicaDatabaseName,
-              replicaTableName);
-      uris.add(location.getPath().toUri());
-    }
-    replicaCatalogListener.deprecatedReplicaLocations(uris);
+    cleanupLocationManager.scheduleLocations();
   }
 
   @Override
@@ -121,26 +96,4 @@ public class FullReplicationReplicaLocationManager implements ReplicaLocationMan
     return replicaPartitionLocation;
   }
 
-  @VisibleForTesting
-  static class CleanUpLocation {
-    private final Path path;
-    private final String pathEventId;
-
-    @VisibleForTesting
-    CleanUpLocation(String pathEventId, Path path) {
-      this.pathEventId = pathEventId;
-      this.path = path;
-    }
-
-    @VisibleForTesting
-    Path getPath() {
-      return path;
-    }
-
-    @VisibleForTesting
-    String getPathEventId() {
-      return pathEventId;
-    }
-
-  }
 }